requirements.txt
colorama==0.4.6
docx2pdf==0.1.8
et-xmlfile==1.1.0
lxml==5.2.2
numpy==2.0.0
openpyxl==3.1.5
pandas==2.2.2
pypdf==4.3.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
pywin32==306
six==1.16.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
